feat:修改文件结构,支持青龙部署。
