+++ README.md
@@ -1,9 +1,9 @@
-# Kuro-autosignin
-python脚本库街区论坛及游戏自动签到
-库街区 游戏及库洛币自动签到
+# Kuro-AutoSignin
+
+自动化每日任务，轻松管理库街区论坛与游戏签到
 
 ## 使用说明
 
-1. **替换个人信息**：请通过抓库街区的包替换脚本中的 `token`、`devcode` 和 `distinct_id`。
+1. **替换个人信息**：请通过抓库街区的包替换脚本中的 `token`、`devcode`、 `roleId`和 `userId`。
 2. **获取 Token**：使用 `tools.py` 来获取你的 `token`(此方法用于不想折腾安卓抓包并且客户端不能再次登录，库街区的token如果新设备登陆了会刷新)；或者直接抓包
-3. **已对接server酱**
+3. **已对接server酱**:在 serverKey 字段中填入Server酱 Key，激活即时通知服务。
