--- serverjiang.py
@@ -1,16 +1,0 @@
-
-import urllib.parse
-import urllib.request
-
-
-def sc_send(text, desp, key=''):
-    postdata = urllib.parse.urlencode({'text': text, 'desp': desp}).encode('utf-8')
-    url = f'https://sctapi.ftqq.com/{key}.send'
-    req = urllib.request.Request(url, data=postdata, method='POST')
-    with urllib.request.urlopen(req) as response:
-        result = response.read().decode('utf-8')
-    return result
-
-
-
-
